cyber: delete examples/tools.py -- the gym ships no agent tools
OpenRange is the gym: procedural worlds, a held-out grader, and the per-episode
EpisodeEnv surface (base_url, solver_root, a sandboxed run, run_tests). Tools are the
agent harness's job. The shipped examples/tools.py reinvented what every harness
already provides (a shell, file read/write/edit, apply_patch), so it's gone.

A consumer brings its own two tiny tools -- a `shell` that forwards to surface["run"]
and a `submit` that writes result.json for the grader -- or points a harness at the
surface (examples/strands_eval.py already uses strands_tools.shell; openrange[strands]
is an existing extra). The few tests that drive an agent define those inline, and
tests/test_byo_tools.py still proves the gym is solvable with nothing shipped.
