+++ tests/test_trl_sandbox_episode.py
@@ -27,7 +27,6 @@ from graphschema import WorldGraph
 from openrange_pack_sdk import Backing, Snapshot
 from openrange_trl import EpisodeEnv, SandboxError
 
-from examples.tools import shell
 from openrange.core.admit import admit
 from openrange.core.episode import EpisodeService
 
@@ -53,8 +52,17 @@ _SSRF_MANIFEST = {
 }
 
 
-# The agent's `shell` is the shipped example tool (examples/tools.py); `submit` is the
-# user's own, recording the answer the held-out grader reads.
+# Both tools are the user's own bring-your-own: `shell` reads the `run` capability the
+# env injects, and `submit` records the answer the held-out grader reads.
+
+
+def shell(surface: Mapping[str, Any], command: str) -> str:
+    """Run a shell command on your machine and return its output.
+
+    Args:
+        command: the shell command to run.
+    """
+    return str(surface["run"](command).output)
 
 
 def submit(surface: Mapping[str, Any], flag: str) -> str:
